+++ backend/transactions/models.py
@@ -1,104 +1,16 @@
 """
-transactions/models.py — Bank, AccountType, Account, Transaction, and Label models.
+transactions/models.py — Label and Transaction models.
 """
 
 from django.db import models
 
-from transactions.constants import HandlerKeys
+from banking.models import (  # noqa: F401  # noqa: F401 — re-exported for any existing imports
+    Account,
+    AccountType,
+    Bank,
+)
+from transactions.constants import HandlerKeys  # noqa: F401 — still used by handlers until Task 3
 from users.models import Household
-
-
-class Bank(models.Model):
-    """
-    Represents a financial institution.
-    Suported banks are sytem determined. Adding, edditing or removing
-    institutions should be performed through migrations.
-    """
-
-    name = models.CharField(max_length=255, unique=True)
-    logo = models.ImageField(upload_to='banks/', blank=True, null=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        db_table = 'banks'
-
-    def __str__(self):
-        return self.name
-
-
-class AccountType(models.Model):
-    """
-    Represents a type of account offered by a bank (e.g. 360 Performance Savings).
-    Maps to a handler in ACCOUNT_HANDLERS via handler_key.
-    Each bank can offer multiple account types, each with a unique handler.
-    """
-
-    HANDLER_CHOICES = [
-        (HandlerKeys.SOFI_SAVINGS, 'SoFi Savings'),
-        (HandlerKeys.SOFI_CHECKING, 'SoFi Checking'),
-        (HandlerKeys.CO_CHECKING, '360 Checking'),
-        (HandlerKeys.CO_SAVINGS, '360 Performance Savings'),
-        (HandlerKeys.CO_QUICKSILVER, 'Quicksilver Credit Card'),
-        (HandlerKeys.WF_CHECKING, 'Checking'),
-        (HandlerKeys.WF_SAVINGS, 'Savings'),
-        (HandlerKeys.CHASE, 'Chase Card'),
-        (HandlerKeys.DISCOVER, 'Discover Card'),
-        (HandlerKeys.AMEX_DELTA, 'Delta SkyMiles Card'),
-    ]
-
-    name = models.CharField(max_length=255)
-    handler_key = models.CharField(
-        max_length=255,
-        unique=True,
-        choices=HANDLER_CHOICES,
-    )
-    bank = models.ForeignKey(
-        Bank,
-        on_delete=models.PROTECT,
-        related_name='account_types',
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        db_table = 'account_types'
-        unique_together = [['bank', 'name']]
-
-    def __str__(self):
-        return f'{self.bank.name} — {self.name}'
-
-    def get_handler(self):
-        """Returns the handler associated with the selected account type."""
-        from transactions.handlers.accounts import ACCOUNT_HANDLERS
-
-        return ACCOUNT_HANDLERS[self.handler_key]
-
-
-class Account(models.Model):
-    """
-    Represents a specific account belonging to a household.
-    Name is user-editable to distinguish multiple accounts of the same type
-    (e.g. "Mario's 360 Savings" and "Partner's 360 Savings").
-    The handler is resolved through account_type.handler_key.
-    """
-
-    name = models.CharField(max_length=255)
-    account_type = models.ForeignKey(AccountType, on_delete=models.PROTECT, related_name='accounts')
-    household = models.ForeignKey(Household, on_delete=models.PROTECT, related_name='accounts')
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        db_table = 'accounts'
-        unique_together = [['household', 'name']]
-
-    def __str__(self):
-        return f'{self.account_type.bank.name} — {self.name}'
-
-    @property
-    def handler_key(self):
-        return self.account_type.handler_key
 
 
 class Label(models.Model):
@@ -162,7 +74,7 @@ class Transaction(models.Model):
     created ones ('manual'). These are treated as distinct entry points but
     share the same dedup scope.
 
-     label links to a Label instance within the same household. When a Label is
+    label links to a Label instance within the same household. When a Label is
     deleted the FK is set to NULL (SET_NULL) so the transaction is preserved but
     simply becomes unlabelled. Earlier versions of this model used a free-text
     label field; that has now been replaced by the structured ForeignKey.
@@ -202,7 +114,11 @@ class Transaction(models.Model):
             'otherwise distort spending/earnings totals.'
         ),
     )
-    account = models.ForeignKey(Account, on_delete=models.PROTECT, related_name='transactions')
+    account = models.ForeignKey(
+        'banking.Account',
+        on_delete=models.PROTECT,
+        related_name='transactions',
+    )
     imported_at = models.DateTimeField(auto_now_add=True)
 
     class Meta:
